+++ src/model/MergeOrganizationsRequest.js
@@ -56,7 +56,7 @@ class MergeOrganizationsRequest {
             }
 
             if (Object.keys(data).length > 0) {
-                obj['extra'] = data;
+                Object.assign(obj, data);
             }
 
         }