+++ src/model/NewDeal.js
@@ -123,7 +123,7 @@ class NewDeal {
             }
 
             if (Object.keys(data).length > 0) {
-                obj['extra'] = data;
+                Object.assign(obj, data);
             }
 
         }