+++ src/model/NewOrganization.js
@@ -78,7 +78,7 @@ class NewOrganization {
             }
 
             if (Object.keys(data).length > 0) {
-                obj['extra'] = data;
+                Object.assign(obj, data);
             }
 
         }