+++ src/model/OrganizationRelationship.js
@@ -69,7 +69,7 @@ class OrganizationRelationship {
             }
 
             if (Object.keys(data).length > 0) {
-                obj['extra'] = data;
+                Object.assign(obj, data);
             }
 
         }